Refactor UserData.
