+++ app_dart/lib/src/request_handlers/rerun_prod_task.dart
@@ -22,7 +22,7 @@ import '../service/datastore.dart';
 import '../service/firestore.dart';
 import '../service/luci_build_service.dart';
 import '../service/luci_build_service/build_tags.dart';
-import '../service/luci_build_service/user_data.dart';
+import '../service/luci_build_service/firestore_task_document_name.dart';
 import '../service/scheduler.dart';
 import '../service/scheduler/ci_yaml_fetcher.dart';
 
@@ -199,7 +199,7 @@ final class RerunProdTask extends ApiRequestHandler<Body> {
     }
 
     // Prepares Firestore task.
-    final documentName = FirestoreTaskDocumentName(
+    final firestoreTask = FirestoreTaskDocumentName(
       commitSha: commitSha,
       taskName: taskName,
       currentAttempt: task.attempts!,
@@ -210,7 +210,7 @@ final class RerunProdTask extends ApiRequestHandler<Body> {
         kDatabase,
         'documents',
         firestore.kTaskCollectionId,
-        '$documentName',
+        firestoreTask.documentName,
       ),
     );
